Utils: Separate into submodules
Split utility functions into several files rather than one very large
(and ever-expanding) file. Import all the functions with `util`'s
`__init__.py` to provide backwards compatibility.
