+++ debian/changelog
@@ -1,3 +1,9 @@
+kano-toolset (2.3.0-2) unstable; urgency=low
+
+  * Separate utils module into submodule.
+
+ -- Team Kano <dev@kano.me>  Wed, 6 Jan 2016 17:15:00 +0100
+
 kano-toolset (2.3.0-1) unstable; urgency=low
 
   * Add decorators module.
